Fix: This guest appears to be already running/loaded.
There is a race condition where iohyve will fail to start a bhyve
instance when there is a running iohyve with the prefix of the same name
as the iohyve you are trying to start.

e.g. (you have `example-other` running):

`iohyve start example`

Will fail with the error message:

```
This guest appears to be already running/loaded.
Use stop, destroy, or forcekill to remove zombie processes.
```

This corrects the pgrep regexs to stop matching past the name.